Refactor bills..... again
After our meeting on 30/10, it turned out that no one is gonna do
bills automation because values of the fees changing all the time.
So now our new approach is this:
- Remove all of the old code related to creating master bills and
automation.
- Bills is now assigned to room_id, and should expect to be created
manually once per month.
- It will include all the fees (water, electricity, etc.) in one
single bill. When creating bill, create an endpoint to know which
room haven't issued a bill yet.
- Split Creating new issued bills & draft bills into 2 different
functions to make status set when "Create" or "Save As Draft" button
is set. Do the same for bills update.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>