Integrate PayOS as a new payment method
VNPay need time to sign in as a merchant and doing some contract work,
which we don't like it. So I found this service called PayOS by CASSO
which provide a personal account with just a citizen ID so it's much
faster to us to do actual payment instead of sandboxing.

Prepare function to create payment with PayOS link, and an endpoint
to receive webhook to know if payment is success or failed.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>